+++ README.md
@@ -3,10 +3,10 @@
 
 #### use
 ```java
-package cn.backpackerxl.jpkit;
+package cn.zzwl.jpkit;
 
-import cn.backpackerxl.jpkit.core.JSON;
-import cn.backpackerxl.jpkit.typeof.*;
+import JSON;
+import cn.zzwl.jpkit.typeof.*;
 import org.junit.Test;
 
 public class JSONTest {
