--- src/main/java/cn/backpackerxl/jpkit/typeof/JBase.java
@@ -1,6 +1,0 @@
-package cn.backpackerxl.jpkit.typeof;
-
-import cn.backpackerxl.jpkit.core.ITypeof;
-
-public abstract class JBase implements ITypeof<Object> {
-}
