+++ src/main/java/cn/zzwl/jpkit/exception/JCharacterException.java
@@ -1,4 +1,4 @@
-package cn.backpackerxl.jpkit.exception;
+package cn.zzwl.jpkit.exception;
 
 public class JCharacterException extends RuntimeException {
     public JCharacterException(String message) {
