+++ src/main/java/cn/zzwl/jpkit/typeof/JBase.java
@@ -1,0 +1,6 @@
+package cn.zzwl.jpkit.typeof;
+
+import cn.zzwl.jpkit.core.ITypeof;
+
+public abstract class JBase implements ITypeof<Object> {
+}
