+++ src/main/java/cn/zzwl/jpkit/typeof/JLong.java
@@ -1,4 +1,4 @@
-package cn.backpackerxl.jpkit.typeof;
+package cn.zzwl.jpkit.typeof;
 
 public class JLong extends JBase {
     private final Long value;
